services/notify_worker/requirements.txt
pika==1.2
pydantic==1.8.2
backoff==1.10.0
Jinja2==3.0.1
psycopg2-binary==2.8.6
